fix: move header files from component folder to layout folder
