+++ src/layout/index.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import Header from '../components/Header';
+import Header from './Header';
 import Footer from './Footer';
 
 const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
